pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["organize_pictures*"]
exclude = ["docs*", "tests*"]

[project]
name = 'OrganizePictures'
version = '1.0.7'
description='A tool for organizing pictures into folders by date.'

authors = [
    {name = "Jacob Truman", email="jacob.truman@gmail.com"},
]

dependencies = [
    "pytz>=2025.2",
    "PyExifTool>=0.5.6",
    "pymediainfo>=7.0.1",
    "ffmpeg-python>=0.2.0",
    "pillow>=11.3.0",
    "pillow_heif>=1.1.0",
    "xmltodict>=1.0.2",
    "dict2xml>=1.7.6",
    "python-magic>=0.4.27",
    "numpy>=2.3.3",
]

requires-python = ">=3.11, <4"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
homepage = "https://github.com/jacobtruman/OrganizePictures"
documentation = "https://github.com/jacobtruman/OrganizePictures"
repository = "https://github.com/jacobtruman/OrganizePictures"

[project.scripts]
organizepictures = 'organize_pictures.scripts.organizepictures:main'
truexif = 'organize_pictures.scripts.truexif:main'
trugiftomp4 = 'organize_pictures.scripts.trugiftomp4:main'
